flyttat koden tillbaka efter testdagen

+++ reglering-v1/PwmTest/PwmTest/src/MathFunctions.c
@@ -16,9 +16,9 @@
    @param radian the angle value in radian
    return degree the angle value in degree
 **/
-double radianToDegree (double radian){
-	double degree;
-	degree = radian*(180/M_PI);
+int radianToDegree (int radian){
+	int degree;
+	degree = (int) radian*(180/M_PI);
 	return degree;
 }
 
@@ -29,11 +29,11 @@ double radianToDegree (double radian){
    @param Yb the y coordinate for B
    return distance the distance between A and B
 **/
-double calculateDistance (double Xa, double Ya, double Xb, double Yb){
-	double distance;
-	double deltaX = Xb - Xa;
-	double deltaY = Yb - Ya;
-	distance = sqrt((deltaX*deltaX) + (deltaY*deltaY));
+int calculateDistance (int Xa, int Ya, int Xb, int Yb){
+	int distance;
+	int deltaX = Xb - Xa;
+	int deltaY = Yb - Ya;
+	distance =(int) sqrt((deltaX*deltaX) + (deltaY*deltaY));
 	return distance;
 }
 
@@ -44,10 +44,10 @@ double calculateDistance (double Xa, double Ya, double Xb, double Yb){
    @param Yb the y coordinate for B
    return azimutAngle the azimuth angle from A to B
 **/
-double calculateAzimuthAngle (double Xa, double Ya, double Xb, double Yb){
-	double azimutAngle = 0;
-	double deltaX = Xb-Xa;
-	double deltaY = Yb-Ya;
+int calculateAzimuthAngle (int Xa, int Ya, int Xb, int Yb){
+	int azimutAngle = 0;
+	int deltaX = Xb-Xa;
+	int deltaY = Yb-Ya;
 	
 	// if A and B has same coordinates
 	if ( (deltaX == 0) && (deltaY == 0) )
@@ -89,17 +89,17 @@ double calculateAzimuthAngle (double Xa, double Ya, double Xb, double Yb){
 		// the angle is between 90 to 180 degree
 		else if ( (deltaX > 0) && (deltaY < 0 ) )
 		{
-			azimutAngle = (radianToDegree (atan (deltaX/deltaY))  + 180);
+			azimutAngle = (radianToDegree ((int)atan (deltaX/deltaY))  + 180);
 		}
 		// the angle is between 180 to 270 degree
 		else if ( (deltaX < 0) && (deltaY < 0 ) )
 		{
-			azimutAngle = (radianToDegree (atan (deltaX/deltaY) ) + 180);
+			azimutAngle = (radianToDegree ((int)atan (deltaX/deltaY) ) + 180);
 		}
 		// the angle is between 270 to 360 degree
 		else if ( (deltaX < 0) && (deltaY > 0 ) )
 		{
-			azimutAngle = (radianToDegree (atan (deltaX/deltaY) ) + 360);
+			azimutAngle = (radianToDegree ((int)atan (deltaX/deltaY) ) + 360);
 		}
 	}
 	
@@ -116,11 +116,11 @@ double calculateAzimuthAngle (double Xa, double Ya, double Xb, double Yb){
    @param Yc the y coordinate for C
    return rotationAngle the rotation angle from A to target C
    **/
-double calcluteRotationAngle (double Xa, double Ya, double Xb, double Yb, double Xc, double Yc){
+int calcluteRotationAngle (int Xa, int Ya, int Xb, int Yb, int Xc, int Yc){
 	
-	double AoB = calculateAzimuthAngle(Xa, Ya, Xb, Yb);
-	double BoC = calculateAzimuthAngle(Xb, Yb, Xc, Yc);
-	double rotationAngle = BoC-AoB;
+	int AoB = calculateAzimuthAngle(Xa, Ya, Xb, Yb);
+	int BoC = calculateAzimuthAngle(Xb, Yb, Xc, Yc);
+	int rotationAngle = BoC-AoB;
 	if (rotationAngle > 180)
 	{
 		rotationAngle =  rotationAngle - 360;
